Return main upstream response when using multiple endpoints
This changes refactors the Proxy remote wrtie handler to do
as much work outside of the handler as possible.

There is a major change in behaviour.
The extra endpoints are now written to async in the background of
the main request. Extra endpoints will be written to with best
effort and logged.

The response code from the thanos backend will be returned to the
client.

This will cause the client to retry on specific errors and will
have a knock on effect of a successfully written request prior
being attempted once again to the extra remote write endpoints.